requirements.txt
watchfiles
pydantic
pydantic-settings
markitdown[all]
docling
Pillow
onnxruntime
tokenizers
sentence-transformers
keybert
plyer
rich
psutil
blinker
sqlmodel
huggingface-hub
numpy